+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v0.4.0
 * Added 'booleanKey' type.
 * Added 'interface' type. Same behavior as object type, except it is non-strict by default.
+* Added 'defaultValue' option to 'optional'.
 * ...
 
 v0.3.3